+++ integrations/airtable/src/index.ts
@@ -15,6 +15,7 @@ import { Base } from "./base";
 import { Webhooks, createWebhookEventSource } from "./webhooks";
 
 export * from "./types";
+export * from "./base";
 
 export type AirtableIntegrationOptions = {
   /** An ID for this client  */
@@ -92,7 +93,7 @@ export class Airtable implements TriggerIntegration {
     if (!this._io) throw new Error("No IO");
     if (!this._connectionKey) throw new Error("No connection key");
 
-    return this._io.runTask<TResult>(
+    return this._io.runTask(
       key,
       (task, io) => {
         if (!this._client) throw new Error("No client");