+++ integrations/replicate/src/types.ts
@@ -1,0 +1,3 @@
+export type CallbackTimeout = { timeoutInSeconds?: number };
+
+export type ReplicateReturnType<T> = Promise<T>;