+++ integrations/stripe/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @trigger.dev/stripe
 
+## 2.2.0
+
+### Patch Changes
+
+- Updated dependencies [975c5f1d]
+  - @trigger.dev/integration-kit@2.2.0
+  - @trigger.dev/sdk@2.2.0
+
+## 2.1.9
+
+### Patch Changes
+
+- 9a187f9e: upgrade zod to 3.22.3
+- Updated dependencies [9a187f9e]
+- Updated dependencies [2e9452ab]
+  - @trigger.dev/sdk@2.1.9
+  - @trigger.dev/integration-kit@2.1.9
+
+## 2.1.8
+
+### Patch Changes
+
+- Updated dependencies [6a992a19]
+- Updated dependencies [ab9e4a98]
+- Updated dependencies [ab9e4a98]
+  - @trigger.dev/sdk@2.1.8
+  - @trigger.dev/integration-kit@2.1.8
+
 ## 2.1.7
 
 ### Patch Changes